# apifox-openapi.yaml
openapi: 3.0.3
info:
  title: Trip-Camp 酒店项目 API
  version: 1.0.0
  description: |
    面向：用户端（Taro/H5/小程序）、管理端（React+AntD）、服务端（Express+MongoDB）。

    约定：
    - Base URL：/api/v1
    - 鉴权：Bearer Token（Authorization: Bearer <token>）
    - 角色：user / merchant / admin
servers:
  - url: http://localhost:3000/api/v1
    description: local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
          example: Token missing or invalid
        details:
          type: object
          additionalProperties: true
      required: [code, message]
    UserPublic:
      type: object
      properties:
        id:
          type: string
          example: 65b8d9f1e8c2f10a1b2c3d4e
        username:
          type: string
          example: alice
        role:
          type: string
          enum: [user, merchant, admin]
          example: user
      required: [id, username, role]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/UserPublic'
      required: [token, user]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          example: alice
        password:
          type: string
          minLength: 6
          example: Password123!
        role:
          type: string
          enum: [user, merchant]
          example: user
      required: [username, password, role]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: alice
        password:
          type: string
          example: Password123!
      required: [username, password]
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: 注册（用户/商户）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              user:
                value: { username: alice, password: Password123!, role: user }
              merchant:
                value: { username: bob_shop, password: Password123!, role: merchant }
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: 用户名已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: 登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example: { username: alice, password: Password123! }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [Auth]
      summary: 获取当前登录用户
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserPublic'
                required: [user]
        '401':
          description: 未登录或 token 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...